[Issue1] 발자취 업로드 이후 인증 화면 진입 안되는 이슈
1. 패키지명 정리
2. viewmodel과 repository 분리 및 정리
