--- app/src/main/java/com/yjdev/goforawalk/LoginUiState.kt
@@ -1,8 +1,0 @@
-package com.yjdev.goforawalk
-
-sealed class LoginUiState {
-    data object Idle : LoginUiState()
-    data object Cancelled : LoginUiState()
-    data object Success : LoginUiState()
-    data class Failure(val error: String) : LoginUiState()
-}
