--- app/src/main/java/com/yjdev/goforawalk/data/ErrorResponse.kt
@@ -1,9 +1,0 @@
-package com.yjdev.goforawalk.data
-
-import com.google.gson.annotations.SerializedName
-
-data class ErrorResponse(
-    @SerializedName("code") val code: String,
-    @SerializedName("message") val message: String,
-    @SerializedName("detailMessage") val detailMessage: String?
-)
