--- app/src/main/java/com/yjdev/goforawalk/data/Feed.kt
@@ -1,10 +1,0 @@
-package com.yjdev.goforawalk.data
-
-data class Feed(
-    val nickname: String,
-    val streak: String,
-    val date: String,
-    val timeAgo: String,
-    val description: String,
-    val imageUrl: String = ""
-)
