--- app/src/main/java/com/yjdev/goforawalk/data/FootStepRequest.kt
@@ -1,5 +1,0 @@
-package com.yjdev.goforawalk.data
-
-data class FootstepResponse(
-    val data: Footstep,
-)
