--- app/src/main/java/com/yjdev/goforawalk/data/NetworkModule.kt
@@ -1,33 +1,0 @@
-package com.yjdev.goforawalk.data
-
-import com.yjdev.goforawalk.ApiService
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-@Module
-@InstallIn(SingletonComponent::class)
-object NetworkModule {
-    val logging = HttpLoggingInterceptor().apply {
-        level = HttpLoggingInterceptor.Level.BODY
-    }
-
-    val client = OkHttpClient.Builder()
-        .addInterceptor(logging)
-        .build()
-
-    @Provides
-    fun provideAuthService(): ApiService {
-        return Retrofit.Builder()
-            .baseUrl("https://api.goforawalk.site/")
-            .client(client)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-            .create(ApiService::class.java)
-    }
-}
