--- app/src/main/java/com/yjdev/goforawalk/data/PostResult.kt
@@ -1,6 +1,0 @@
-package com.yjdev.goforawalk.data
-
-sealed class PostResult {
-    data class Success(val body: Any?) : PostResult()
-    data class Failure(val errorMsg: String) : PostResult()
-}
