--- app/src/main/java/com/yjdev/goforawalk/data/Screen.kt
@@ -1,10 +1,0 @@
-package com.yjdev.goforawalk.data
-
-import com.yjdev.goforawalk.R
-
-sealed class Screen(val route: String, val icon: Int, val label: String) {
-    object Home : Screen("home", R.drawable.ic_home_24, "홈")
-    object Certify : Screen("certify", R.drawable.ic_certify_24, "인증")
-    object Profile : Screen("profile", R.drawable.ic_profile_24, "프로필")
-    object Settings : Screen("settings", R.drawable.ic_profile_24, "설정")
-}
