+++ app/src/main/java/com/yjdev/goforawalk/data/local/TokenManager.kt
@@ -1,0 +1,36 @@
+package com.yjdev.goforawalk.data.local
+
+import android.content.Context
+import android.content.SharedPreferences
+import androidx.security.crypto.EncryptedSharedPreferences
+import androidx.security.crypto.MasterKey
+import dagger.hilt.android.qualifiers.ApplicationContext
+import javax.inject.Inject
+
+class TokenManager @Inject constructor(
+    @ApplicationContext context: Context
+) {
+    private val masterKey = MasterKey.Builder(context)
+        .setKeyScheme(MasterKey.KeyScheme.AES256_GCM)
+        .build()
+
+    private val sharedPreferences: SharedPreferences = EncryptedSharedPreferences.create(
+        context,
+        "secure_prefs",
+        masterKey,
+        EncryptedSharedPreferences.PrefKeyEncryptionScheme.AES256_SIV,
+        EncryptedSharedPreferences.PrefValueEncryptionScheme.AES256_GCM
+    )
+
+    fun saveToken(token: String) {
+        sharedPreferences.edit().putString("access_token", token).apply()
+    }
+
+    fun getToken(): String? {
+        return sharedPreferences.getString("access_token", null)
+    }
+
+    fun clearToken() {
+        sharedPreferences.edit().remove("access_token").apply()
+    }
+}
