+++ app/src/main/java/com/yjdev/goforawalk/data/model/ErrorResponse.kt
@@ -1,0 +1,9 @@
+package com.yjdev.goforawalk.data.model
+
+import com.google.gson.annotations.SerializedName
+
+data class ErrorResponse(
+    @SerializedName("code") val code: String,
+    @SerializedName("message") val message: String,
+    @SerializedName("detailMessage") val detailMessage: String?
+)
