+++ app/src/main/java/com/yjdev/goforawalk/data/model/FootStepRequest.kt
@@ -1,0 +1,5 @@
+package com.yjdev.goforawalk.data.model
+
+data class FootstepResponse(
+    val data: Footstep,
+)
