+++ app/src/main/java/com/yjdev/goforawalk/data/model/Footstep.kt
@@ -1,0 +1,19 @@
+package com.yjdev.goforawalk.data.model
+
+data class FootstepsResponse(
+    val data: FootstepData
+)
+
+data class FootstepData(
+    val footsteps: List<Footstep>
+)
+
+data class Footstep(
+    val userId: Int,
+    val userNickname: String,
+    val footstepId: Int,
+    val date: String,
+    val imageUrl: String,
+    val content: String?,
+    val createdAt: String
+)
