+++ app/src/main/java/com/yjdev/goforawalk/domain/model/PostResult.kt
@@ -1,0 +1,6 @@
+package com.yjdev.goforawalk.domain.model
+
+sealed class PostResult {
+    data class Success(val body: Any?) : PostResult()
+    data class Failure(val errorMsg: String) : PostResult()
+}
