+++ app/src/main/java/com/yjdev/goforawalk/presentation/ui/theme/Color.kt
@@ -1,0 +1,20 @@
+package com.yjdev.goforawalk.presentation.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+val Purple80 = Color(0xFFD0BCFF)
+val PurpleGrey80 = Color(0xFFCCC2DC)
+val Pink80 = Color(0xFFEFB8C8)
+
+val Purple40 = Color(0xFF6650a4)
+val PurpleGrey40 = Color(0xFF625b71)
+val Pink40 = Color(0xFF7D5260)
+
+val Gray757990 = Color(0xFF757990)
+val GrayDDDDDD = Color(0xFFDDDDDD)
+val GrayD3D3D3 = Color(0xFFD3D3D3)
+val YellowFEE500 = Color(0xFFFEE500)
+val Green8BC34A = Color(0xFF8BC34A)
+val Green8AA76D = Color(0xFF8AA76D)
+
+val Black3A1D1D = Color(0xFF3A1D1D)
