--- app/src/main/java/com/yjdev/goforawalk/repository/FootStepRepository.kt
@@ -1,10 +1,0 @@
-package com.yjdev.goforawalk.repository
-
-import com.yjdev.goforawalk.ApiService
-import javax.inject.Inject
-
-class FootstepRepository @Inject constructor(
-    private val api: ApiService
-) {
-    suspend fun deleteFootstep(auth: String, id: Int) = api.deleteFootstep(auth, id)
-}
