--- app/src/main/java/com/yjdev/goforawalk/repository/LoginRepository.kt
@@ -1,23 +1,0 @@
-package com.yjdev.goforawalk.repository
-
-import android.content.Context
-import android.util.Log
-import com.kakao.sdk.auth.model.OAuthToken
-import com.kakao.sdk.user.UserApiClient
-import javax.inject.Inject
-
-class LoginRepository @Inject constructor() {
-    suspend fun loginWithKakao(context: Context, callback: (OAuthToken?, Throwable?) -> Unit) {
-        if (UserApiClient.instance.isKakaoTalkLoginAvailable(context)) {
-            UserApiClient.instance.loginWithKakaoTalk(context) { token, error ->
-                if (error != null) {
-                    Log.e("LoginRepository", "카카오톡으로 로그인 실패", error)
-                } else {
-                    callback(token, null)
-                }
-            }
-        } else {
-            UserApiClient.instance.loginWithKakaoAccount(context, callback = callback)
-        }
-    }
-}
