--- app/src/main/java/com/yjdev/goforawalk/ui/screen/CameraScreen().kt
@@ -1,102 +1,0 @@
-package com.yjdev.goforawalk.ui.screen
-
-import android.net.Uri
-import android.view.Surface
-import androidx.camera.core.CameraSelector
-import androidx.camera.core.ImageCapture
-import androidx.camera.core.ImageCaptureException
-import androidx.camera.core.Preview
-import androidx.camera.lifecycle.ProcessCameraProvider
-import androidx.camera.view.PreviewView
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.viewinterop.AndroidView
-import androidx.core.content.ContextCompat
-import androidx.lifecycle.LifecycleOwner
-import java.io.File
-
-@Composable
-fun CameraScreen(
-    onImageCaptured: (Uri) -> Unit,
-    onError: (ImageCaptureException) -> Unit
-) {
-    val context = LocalContext.current
-    val imageCapture: ImageCapture = remember {
-        ImageCapture.Builder()
-            .setTargetRotation(Surface.ROTATION_0)
-            .build()
-    }
-
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(Color.Black)
-    ) {
-        AndroidView(
-            factory = { ctx ->
-                val previewView = PreviewView(ctx)
-                val cameraProviderFuture = ProcessCameraProvider.getInstance(ctx)
-
-                cameraProviderFuture.addListener({
-                    val cameraProvider = cameraProviderFuture.get()
-                    val preview = Preview.Builder().build().also {
-                        it.setSurfaceProvider(previewView.surfaceProvider)
-                    }
-
-                    val cameraSelector = CameraSelector.DEFAULT_BACK_CAMERA
-
-                    cameraProvider.unbindAll()
-                    cameraProvider.bindToLifecycle(
-                        context as LifecycleOwner, cameraSelector, preview, imageCapture
-                    )
-                }, ContextCompat.getMainExecutor(ctx))
-
-                previewView
-            },
-            modifier = Modifier
-                .height(120.dp)
-                .align(Alignment.Center)
-        )
-
-        Box(
-            modifier = Modifier
-                .align(Alignment.BottomCenter)
-                .padding(bottom = 20.dp)
-                .background(Color.White, shape = CircleShape)
-                .size(64.dp)
-                .clickable {
-                    val photoFile = File(context.cacheDir, "camera_image.jpg")
-                    val outputOptions = ImageCapture.OutputFileOptions
-                        .Builder(photoFile)
-                        .build()
-
-                    imageCapture.takePicture(
-                        outputOptions,
-                        ContextCompat.getMainExecutor(context),
-                        object : ImageCapture.OnImageSavedCallback {
-                            override fun onImageSaved(output: ImageCapture.OutputFileResults) {
-                                onImageCaptured(Uri.fromFile(photoFile))
-                            }
-
-                            override fun onError(exception: ImageCaptureException) {
-                                onError(exception)
-                            }
-                        }
-                    )
-                }
-        )
-    }
-}
