--- app/src/main/java/com/yjdev/goforawalk/ui/screen/SettingsScreen.kt
@@ -1,59 +1,0 @@
-package com.yjdev.goforawalk.ui.screen
-
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ArrowBack
-import androidx.compose.material3.Icon
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-
-@Composable
-fun SettingsScreen(onBack: () -> Unit) {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(16.dp)
-    ) {
-        Row(
-            verticalAlignment = Alignment.CenterVertically,
-            modifier = Modifier.clickable { onBack() }
-        ) {
-            Icon(
-                imageVector = Icons.Default.ArrowBack,
-                contentDescription = "뒤로가기",
-                modifier = Modifier.size(24.dp)
-            )
-            Spacer(modifier = Modifier.width(8.dp))
-            Text(
-                text = "설정",
-                fontWeight = FontWeight.Bold,
-                fontSize = 20.sp
-            )
-        }
-        Spacer(modifier = Modifier.height(24.dp))
-
-        Text(
-            text = "회원 탈퇴",
-            fontWeight = FontWeight.Bold,
-            fontSize = 16.sp,
-            color = Color.DarkGray,
-            modifier = Modifier
-                .clickable { /* 회원 탈퇴 기능 */ }
-                .padding(vertical = 8.dp)
-        )
-    }
-}
