+++ gradle/libs.versions.toml
@@ -19,6 +19,7 @@ material3 = "1.3.2"
 coil = "2.4.0"
 camera = "1.2.0"
 accompanist = "0.35.0-alpha"
+okhttp3 = "4.11.0"
 
 [libraries]
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
@@ -49,7 +50,10 @@ androidx-camera = { module = "androidx.camera:camera-camera2", version.ref = "ca
 androidx-camera-lifecycle = { module = "androidx.camera:camera-lifecycle", version.ref = "camera" }
 androidx-camera-view = { module = "androidx.camera:camera-view", version.ref = "camera" }
 androidx-camera-extensions = { module = "androidx.camera:camera-extensions", version.ref = "camera" }
-google-accompanist = {module = "com.google.accompanist:accompanist-permissions",version.ref = "accompanist"}
+google-accompanist = { module = "com.google.accompanist:accompanist-permissions", version.ref = "accompanist" }
+okhttp3 = { module = "com.squareup.okhttp3:okhttp", version.ref = "okhttp3" }
+okhttp3-interceptor = { module = "com.squareup.okhttp3:logging-interceptor", version.ref = "okhttp3" }
+
 
 [plugins]
 android-application = { id = "com.android.application", version.ref = "agp" }
